rune/libenclave/attestation: move sgx and registration packages into the internal
Fixes: #613
