+++ rune/libenclave/attestation/challenger.go
@@ -2,8 +2,8 @@ package attestation // import "github.com/inclavare-containers/rune/libenclave/a
 
 import (
 	"fmt"
-	"github.com/inclavare-containers/rune/libenclave/attestation/registration"
-	_ "github.com/inclavare-containers/rune/libenclave/attestation/sgx/challenger" // for the registration of sgx challengers
+	"github.com/inclavare-containers/rune/libenclave/attestation/internal/registration"
+	_ "github.com/inclavare-containers/rune/libenclave/attestation/internal/sgx/challenger" // for the registration of sgx challengers
 )
 
 type Challenger interface {
