+++ rune/libenclave/attestation/internal/sgx/challenger/challenger-epid.go
@@ -1,9 +1,9 @@
-package sgx_challenger // import "github.com/inclavare-containers/rune/libenclave/attestation/sgx/challenger"
+package sgx_challenger // import "github.com/inclavare-containers/rune/libenclave/attestation/internal/sgx/challenger"
 
 import (
 	"fmt"
-	"github.com/inclavare-containers/rune/libenclave/attestation/registration"
-	"github.com/inclavare-containers/rune/libenclave/attestation/sgx/ias"
+	"github.com/inclavare-containers/rune/libenclave/attestation/internal/registration"
+	"github.com/inclavare-containers/rune/libenclave/attestation/internal/sgx/ias"
 	"github.com/inclavare-containers/rune/libenclave/intelsgx"
 	"github.com/sirupsen/logrus"
 )
