+++ rune/libenclave/attestation/internal/sgx/ias/ias.go
@@ -1,4 +1,4 @@
-package ias // import "github.com/inclavare-containers/rune/libenclave/attestation/sgx/ias"
+package ias // import "github.com/inclavare-containers/rune/libenclave/attestation/sgx/internal/ias"
 
 import (
 	"bytes"
